requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
easydict==1.9
fasttext==0.9.1
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybind11==2.4.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytorch-ssim==0.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
tensorboard==2.0.1
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
torch==1.3.1
torchfile==0.1.0
torchtext==0.4.0
torchvision==0.4.2
tqdm==4.38.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
